+++ xclim/sdba/measures.py
@@ -9,21 +9,26 @@ through statistical properties or directly. This framework for the diagnostic te
 """
 from __future__ import annotations
 
+from typing import Sequence
+
 import numpy as np
 import xarray as xr
 
-from xclim import sdba
 from xclim.core.indicator import Indicator, base_registry
 from xclim.core.units import convert_units_to, ensure_delta, units2pint
 from xclim.core.utils import InputKind
+
+from .base import Grouper
+from .utils import _pairwise_spearman
 
 
 class StatisticalMeasure(Indicator):
     """Base indicator class for statistical measures used when validating bias-adjusted outputs.
 
-    Statistical measures either use input data where the time dimension was reduced, or they combine
-    the reduction with the measure. They usually take two arrays as input: "sim" and "ref", "sim" being
-    measured against "ref".
+    Statistical measures use input data where the time dimension was reduced, usually by the computation
+    of a :py:class:`xclim.sdba.properties.StatisticalProperty` instance.  They usually take two arrays
+    as input: "sim" and "ref", "sim" being measured against "ref". The two arrays must have identical
+    coordinates on their common dimensions.
 
     Statistical measures are generally unit-generic. If the inputs have different units, "sim" is converted
     to match "ref".
@@ -47,6 +52,15 @@ class StatisticalMeasure(Indicator):
         # Convert grouping and check if allowed:
         sim = das["sim"]
         ref = das["ref"]
+
+        # Check if common coordinates are identical.
+        newsim, newref = xr.broadcast(sim, ref)
+        for dim in set(sim.dims).union(ref.dims):
+            if [sim[dim].size, ref[dim].size] != [newsim[dim].size, newref[dim].size]:
+                raise ValueError(
+                    f"Common dimension {dim} has different coordinates between ref and sim."
+                )
+
         units_sim = units2pint(sim)
         units_ref = units2pint(ref)
 
@@ -56,7 +70,91 @@ class StatisticalMeasure(Indicator):
         return das, params
 
 
+class StatisticalPropertyMeasure(Indicator):
+    """Base indicator class for statistical properties that include the comparison measure, used when validating bias-adjusted outputs.
+
+    StatisticalPropertyMeasure objects combine the functionalities of
+    :py:class:`xclim.sdba.properties.StatisticalProperty` and
+    :py:class:`xclim.sdba.properties.StatisticalMeasure`.
+
+    Statistical properties usually reduce the time dimension and sometimes more dimensions
+    (for example in spatial properties), sometimes adding a grouping dimension according to
+    the passed value of `group` (e.g.: group='time.month' means the loss of the time dimension
+    and the addition of a month one).
+
+    Statistical measures usually take two arrays as input: "sim" and "ref", "sim" being measured against "ref".
+
+    Statistical property-measures are generally unit-generic. If the inputs have different units,
+    "sim" is converted to match "ref".
+    """
+
+    aspect = None
+    """The aspect the statistical property studies: marginal, temporal, multivariate or spatial."""
+
+    allowed_groups = None
+    """A list of allowed groupings. A subset of dayofyear, week, month, season or group.
+    The latter stands for no temporal grouping."""
+
+    realm = "generic"
+
+    @classmethod
+    def _ensure_correct_parameters(cls, parameters):
+        inputs = {k for k, p in parameters.items() if p.kind == InputKind.VARIABLE}
+        if not inputs.issuperset({"sim", "ref"}):
+            raise ValueError(
+                f"{cls.__name__} requires 'sim' and 'ref' as inputs. Got {inputs}."
+            )
+
+        if "group" not in parameters:
+            raise ValueError(
+                f"{cls.__name__} require a 'group' argument, use the base Indicator"
+                " class if your computation doesn't perform any regrouping."
+            )
+
+        return super()._ensure_correct_parameters(parameters)
+
+    def _preprocess_and_checks(self, das, params):
+        """Perform parent's checks and also check convert units so that sim matches ref."""
+        das, params = super()._preprocess_and_checks(das, params)
+
+        # Convert grouping and check if allowed:
+        if isinstance(params["group"], str):
+            params["group"] = Grouper(params["group"])
+
+        if (
+            self.allowed_groups is not None
+            and params["group"].prop not in self.allowed_groups
+        ):
+            raise ValueError(
+                f"Grouping period {params['group'].prop_name} is not allowed for property "
+                f"{self.identifier} (needs something in "
+                f"{list(map(lambda g: '<dim>.' + g.replace('group', ''), self.allowed_groups))})."
+            )
+
+        # Convert grouping and check if allowed:
+        sim = das["sim"]
+        ref = das["ref"]
+        units_sim = units2pint(sim)
+        units_ref = units2pint(ref)
+
+        if units_sim != units_ref:
+            das["sim"] = convert_units_to(sim, ref)
+
+        return das, params
+
+    def _postprocess(self, outs, das, params):
+        """Squeeze `group` dim if needed."""
+        outs = super()._postprocess(outs, das, params)
+
+        for i in range(len(outs)):
+            if "group" in outs[i].dims:
+                outs[i] = outs[i].squeeze("group", drop=True)
+
+        return outs
+
+
 base_registry["StatisticalMeasure"] = StatisticalMeasure
+base_registry["StatisticalPropertyMeasure"] = StatisticalPropertyMeasure
 
 
 def _bias(sim: xr.DataArray, ref: xr.DataArray) -> xr.DataArray:
@@ -166,7 +264,9 @@ def _ratio(sim: xr.DataArray, ref: xr.DataArray) -> xr.DataArray:
 ratio = StatisticalMeasure(identifier="ratio", compute=_ratio, units="")
 
 
-def _rmse(sim: xr.DataArray, ref: xr.DataArray) -> xr.DataArray:
+def _rmse(
+    sim: xr.DataArray, ref: xr.DataArray, group: str | Grouper = "time"
+) -> xr.DataArray:
     """Root mean square error.
 
     The root mean square error on the time dimension between the simulation and the reference.
@@ -174,9 +274,12 @@ def _rmse(sim: xr.DataArray, ref: xr.DataArray) -> xr.DataArray:
     Parameters
     ----------
     sim : xr.DataArray
-      Data from the simulation (a time-series for each grid-point)
+        Data from the simulation (a time-series for each grid-point)
     ref : xr.DataArray
-      Data from the reference (observations) (a time-series for each grid-point)
+        Data from the reference (observations) (a time-series for each grid-point)
+    group: str
+        Compute the property and measure for each temporal groups individually.
+        Currently not implemented.
 
     Returns
     -------
@@ -197,10 +300,18 @@ def _rmse(sim: xr.DataArray, ref: xr.DataArray) -> xr.DataArray:
     return out.assign_attrs(units=ensure_delta(ref.units))
 
 
-rmse = StatisticalMeasure(identifier="rmse", compute=_rmse)
+rmse = StatisticalPropertyMeasure(
+    identifier="rmse",
+    aspect="temporal",
+    compute=_rmse,
+    allowed_groups=["group"],
+    cell_methods="time: mean",
+)
 
 
-def _mae(sim: xr.DataArray, ref: xr.DataArray) -> xr.DataArray:
+def _mae(
+    sim: xr.DataArray, ref: xr.DataArray, group: str | Grouper = "time"
+) -> xr.DataArray:
     """Mean absolute error.
 
     The mean absolute error on the time dimension between the simulation and the reference.
@@ -208,9 +319,12 @@ def _mae(sim: xr.DataArray, ref: xr.DataArray) -> xr.DataArray:
     Parameters
     ----------
     sim : xr.DataArray
-      data from the simulation (a time-series for each grid-point)
+        data from the simulation (a time-series for each grid-point)
     ref : xr.DataArray
-      data from the reference (observations) (a time-series for each grid-point)
+        data from the reference (observations) (a time-series for each grid-point)
+    group : str
+        Compute the property and measure for each temporal groups individually.
+        Currently not implemented.
 
     Returns
     -------
@@ -231,11 +345,20 @@ def _mae(sim: xr.DataArray, ref: xr.DataArray) -> xr.DataArray:
     return out.assign_attrs(units=ensure_delta(ref.units))
 
 
-mae = StatisticalMeasure(identifier="mae", compute=_mae)
+mae = StatisticalPropertyMeasure(
+    identifier="mae",
+    aspect="temporal",
+    compute=_mae,
+    allowed_groups=["group"],
+    cell_methods="time: mean",
+)
 
 
 def _annual_cycle_correlation(
-    sim: xr.DataArray, ref: xr.DataArray, window: int = 15
+    sim: xr.DataArray,
+    ref: xr.DataArray,
+    window: int = 15,
+    group: str | Grouper = "time",
 ) -> xr.DataArray:
     """Annual cycle correlation.
 
@@ -246,12 +369,15 @@ def _annual_cycle_correlation(
     Parameters
     ----------
     sim : xr.DataArray
-      data from the simulation (a time-series for each grid-point)
+        data from the simulation (a time-series for each grid-point)
     ref : xr.DataArray
-      data from the reference (observations) (a time-series for each grid-point)
-    window: int
-      Size of window around each day of year around which to take the mean.
-      E.g. If window=31, Jan 1st is averaged over from December 17th to January 16th.
+        data from the reference (observations) (a time-series for each grid-point)
+    window : int
+        Size of window around each day of year around which to take the mean.
+        E.g. If window=31, Jan 1st is averaged over from December 17th to January 16th.
+    group : str
+        Compute the property and measure for each temporal groups individually.
+        Currently not implemented.
 
     Returns
     -------
@@ -259,7 +385,7 @@ def _annual_cycle_correlation(
       Annual cycle correlation
     """
     # group by day-of-year and window around each doy
-    grouper_test = sdba.base.Grouper("time.dayofyear", window=window)
+    grouper_test = Grouper("time.dayofyear", window=window)
     # for each day, mean over X day window and over all years to create a smooth avg annual cycle
     sim_annual_cycle = grouper_test.apply("mean", sim)
     ref_annual_cycle = grouper_test.apply("mean", ref)
@@ -267,6 +393,52 @@ def _annual_cycle_correlation(
     return out.assign_attrs(units="")
 
 
-annual_cycle_correlation = StatisticalMeasure(
-    identifier="annual_cycle_correlation", compute=_annual_cycle_correlation
+annual_cycle_correlation = StatisticalPropertyMeasure(
+    identifier="annual_cycle_correlation",
+    aspect="temporal",
+    compute=_annual_cycle_correlation,
+    allowed_groups=["group"],
+)
+
+
+def _scorr(
+    sim: xr.DataArray,
+    ref: xr.DataArray,
+    *,
+    dims: Sequence | None = None,
+    group: str | Grouper = "time",
+):
+    """Spatial correllogram.
+
+    Compute the inter-site correlations of each array, compute the difference in correlations and sum.
+    Taken from Vrac (2018). The spatial and temporal dimensions are reduced.
+
+    Parameters
+    ----------
+    sim : xr.DataArray
+        data from the simulation (a time-series for each grid-point)
+    ref : xr.DataArray
+        data from the reference (observations) (a time-series for each grid-point)
+    dims : sequence of strings, optional
+        Name of the spatial dimensions. If None (default), all dimensions except 'time' are used.
+    group : str
+        Compute the property and measure for each temporal groups individually.
+        Currently not implemented.
+
+    Returns
+    -------
+    xr.DataArray, [dimensionless]
+      Sum of the inter-site correlation differences.
+    """
+    if dims is None:
+        dims = [d for d in sim.dims if d != "time"]
+
+    refcorr = _pairwise_spearman(ref, dims)
+    simcorr = _pairwise_spearman(sim, dims)
+    S_corr = (simcorr - refcorr).sum(["_spatial", "_spatial2"])
+    return S_corr.assign_attrs(units="")
+
+
+scorr = StatisticalPropertyMeasure(
+    identifier="Scorr", aspect="spatial", compute=_scorr, allowed_groups=["group"]
 )